requirements.txt
# AnuRAG Requirements
# Core AI packages â€” Gemini (embeddings/vision) + Claude (optional LLM)
google-generativeai>=0.8.0
google-ai-generativelanguage>=0.6.0
anthropic>=0.45.0  # For Claude support (pip install anthropic)

# LangChain ecosystem
langchain>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0
langchain-chroma>=0.1.0

# Vector stores and embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4

# PDF processing
pypdf>=3.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
unstructured>=0.10.0
PyMuPDF>=1.22.0
pillow>=10.0.0

# OCR and image processing
pytesseract>=0.3.10
rapidocr-onnxruntime
opencv-python>=4.8.1

# NLP and tokenization
tiktoken>=0.5.0
nltk>=3.8.0

# Search and reranking
cohere>=5.0.0
elasticsearch>=8.0.0

# Utilities
python-dotenv>=0.19.0
pydantic>=2.6.1
numpy>=1.26.0
pandas>=1.5.0
tqdm>=4.65.0
tenacity>=8.2.0
requests>=2.31.0

# Visualization
matplotlib>=3.7.0

# Web scraping
beautifulsoup4>=4.12.0
selenium>=4.10.0
webdriver-manager>=4.0.0
lxml>=4.9.0
bs4

# Deep learning (for advanced features)
torch>=2.0.0
transformers>=4.30.0

# Rich console output
rich>=13.0.0

# Table extraction
tabula-py>=2.7.0
openpyxl>=3.1.0
